Structural changes and settting up for future updates #major
This is the first part of a restructure - setting the ground work with only the branch-name action for simplicity and to try and reduce the size of the pr!

The larger change is to ... 

Update semver action to allow usage on any github event. Currently the action relies on github.$x values which are not present in every trigger (workflow_dispatch & schedule)

Update semver action to provide functionality for sirius projects; this is a flag so if there are no code changes since the last releas, dont increment the semver. This will be optional behaviour

Reduce number of actions; the various elements of creating a semver tag do not need to be exposed and are not used directly, so merge into one action

Improve performance. Most calls to the action via uses trigger a build of the binaries; see what is included in the uses call and see if it includes attached files. If it does, then we can attach binaries to the release and make that faster

Allow actions to be run outside of pull_request and push events. Currently, some actions are tied to values only found in those events.