Adding semver action #minor (#73)
Adding in semver action, based on the previous version.

The path, input & output variables have changed. **The new variable for the semver tag this action creates is `tag` (instead of `created_tag`)** - as it is not always created. 

See the README in the action folder for more details / examples.

This also adds a new `release` action to as well so we are not reliant on 3rd party for that.

This is part of a restructure with a general aim to .. 
- Update semver action to allow usage on any github event. 
- Update semver action to provide method to return existing semver
- Reduce number of actions
- Improve performance. 

See [v4 relerase](https://github.com/ministryofjustice/opg-github-actions/releases/tag/v4.0.0) for more info