Fix to push event handling #minor (#77)
From the [last merge workflow](https://github.com/ministryofjustice/opg-github-actions/actions/runs/16798507933) can see that the issue is the states its comparing .. its looking for commits between `main...main` - so finds nothing.

This changes that, so instead of using the default branch as the comparison point it uses the last release. If there is no last release then it will use the default branch. Also added handling for the `push` event file, so it should also find commits in there.

We don't remove duplicate commits, but the semver bump is not additive, so even if it finds 3 majors, it will only bump it once.